Phase 4: lift route-chunk recovery out of App.tsx
App.tsx opened with two pieces that have nothing to do with rendering the
app: the error boundary that catches a lazy route chunk's import()
rejecting, and the URL reader that canonicalises the address before
parsing it. Both sat above the imports, which is also why the import
block was split in two.

- appRouteRecovery.ts  RouteChunkBoundary + readLocationRoute

A .ts rather than the .tsx the shape suggests: RouteChunkBoundary renders
its props and contains no JSX, and react-refresh/only-export-components
would flag the helper next to the component in a .tsx under
--max-warnings 0.

Both carried /** */ blocks; rewritten as one-line notes so the new file
starts at zero comment-prose debt. App.tsx 1312 -> 1275.
